# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/setores": {
      "post": {
        "operationId": "SetorController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSetorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Setores"
        ]
      },
      "get": {
        "operationId": "SetorController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "default": "",
              "example": "tecnologia",
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "id",
              "example": "nome",
              "type": "string"
            }
          },
          {
            "name": "ascending",
            "required": false,
            "in": "query",
            "schema": {
              "default": true,
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Setores"
        ]
      }
    },
    "/setores/{id}": {
      "get": {
        "operationId": "SetorController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Setores"
        ]
      },
      "patch": {
        "operationId": "SetorController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSetorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Setores"
        ]
      },
      "delete": {
        "operationId": "SetorController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Setores"
        ]
      }
    }
  },
  "info": {
    "title": "Cadastro de Controladores",
    "description": "API para gerenciamento de controladores, encarregados e grupo econômico conforme LGPD",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateSetorDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Tecnologia"
          }
        },
        "required": [
          "nome"
        ]
      },
      "UpdateSetorDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          }
        }
      }
    }
  }
}